requirements.txt
absl-py==0.11.0
amqp==2.5.1
asn1crypto==0.24.0
astor==0.8.1
atomicwrites==1.3.0
attrs==19.1.0
axe-selenium-python==2.1.6
beautifulsoup4==4.8.0
billiard==3.6.1.0
blis==0.4.1
boto==2.49.0
boto3==1.9.233
botocore==1.12.233
browsermob-proxy==0.8.0
cached-property==1.5.2
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
dataclasses==0.6
decorator==4.4.0
dill==0.3.3
dj-database-url==0.5.0
Django==2.2.19
django-axes==5.26.0
django-celery-results==1.1.2
django-cors-headers==3.5.0
django-filter==2.3.0
django-vault-helpers==0.8.1
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
drf-yasg==1.20.0
docutils==0.15.2
easyocr==1.3.2
fpdf2==2.0.3
gast==0.4.0
gensim==3.8.0
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.4
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.34.0
gunicorn==19.9.0
h5py==3.1.0
hvac==0.10.14
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
imutils==0.5.2
jira==3.0.1
jmespath==0.9.4
joblib==1.1.0
jsonschema==2.6.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
kombu==4.6.4
lxml==4.3.4
Markdown==3.3.3
matplotlib==3.1.1
more-itertools==7.2.0
multiprocess==0.70.11.1
murmurhash==1.0.2
networkx==2.3
nltk==3.4.4
numpy==1.16.4
opencv-contrib-python==4.1.0.25
packaging==19.2
pandas==1.1.5
pathlib==1.0.1
Pillow==6.1.0
plac==0.9.6
pluggy==0.13.0
portalocker==2.3.0
preshed==3.0.2
protobuf==3.14.0
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
pylev==1.4.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.0
pytesseract==0.2.9
pytest==5.1.3
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.3.1
requests==2.22.0
retry==0.9.2
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==1.0.1
scipy==1.3.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
smmap==3.0.4
smmap2==3.0.1
soupsieve==1.9.2
spacy==2.2.1
spacy-lookups-data==0.1.0
sqlparse==0.3.0
srsly==0.1.0
sentry-sdk==0.10.2
symspellpy==6.7.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tesseract==0.1.3
thinc==7.1.1
threadpoolctl==3.0.0
tqdm==4.36.1
urllib3==1.25.3
vine==1.3.0
wasabi==0.2.2
wcwidth==0.1.7
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
wrapt-timeout-decorator==1.3.1
xgboost==1.5.0
xlrd==1.2.0
XlsxWriter==1.3.8
zipp==0.6.0
